+++ src/Components/Form/Part1.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Grid } from '@material-ui/core';
 import { makeStyles, createStyles } from '@material-ui/core/styles';
-import Button from '../../Components/Button';
+import Button from '../Button';
 import { AnimationContainer } from './styles';
 
 const useStyles = makeStyles(() =>
@@ -21,17 +21,16 @@ const useStyles = makeStyles(() =>
 
 interface Props {
   changePageLeft: () => void;
-  changePageRight: () => void;
 }
 
-const App: React.FC<Props> = ({ changePageLeft, changePageRight }) => {
+const App: React.FC<Props> = ({ changePageLeft }) => {
   const classes = useStyles();
   return (
     <AnimationContainer>
       <div className={classes.text}>
         <p>
           A seguir você encontrará perguntas que te farão refletir sobre seus
-          próprios sentimentos. Parte 5!
+          próprios sentimentos.
         </p>
         <p>É muito importante que você responda com calma e honestidade.</p>
         <p className={classes.p}>
@@ -46,11 +45,8 @@ const App: React.FC<Props> = ({ changePageLeft, changePageRight }) => {
         alignItems="center"
         className={classes.div}
       >
-        <Grid item onClick={changePageRight}>
-          <Button type="submit" text="voltar" />
-        </Grid>
         <Grid item onClick={changePageLeft}>
-          <Button type="button" text="próximo" />
+          <Button type="button" text="Vamos lá" />
         </Grid>
       </Grid>
     </AnimationContainer>
