+++ src/Components/Form/Part5.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Grid } from '@material-ui/core';
 import { makeStyles, createStyles } from '@material-ui/core/styles';
-import Button from '../../Components/Button';
+import Button from '../Button';
 import { AnimationContainer } from './styles';
 
 const useStyles = makeStyles(() =>
@@ -21,16 +21,17 @@ const useStyles = makeStyles(() =>
 
 interface Props {
   changePageLeft: () => void;
+  changePageRight: () => void;
 }
 
-const App: React.FC<Props> = ({ changePageLeft }) => {
+const App: React.FC<Props> = ({ changePageLeft, changePageRight }) => {
   const classes = useStyles();
   return (
     <AnimationContainer>
       <div className={classes.text}>
         <p>
           A seguir você encontrará perguntas que te farão refletir sobre seus
-          próprios sentimentos.
+          próprios sentimentos. Parte 5!
         </p>
         <p>É muito importante que você responda com calma e honestidade.</p>
         <p className={classes.p}>
@@ -45,8 +46,11 @@ const App: React.FC<Props> = ({ changePageLeft }) => {
         alignItems="center"
         className={classes.div}
       >
+        <Grid item onClick={changePageRight}>
+          <Button type="submit" text="voltar" />
+        </Grid>
         <Grid item onClick={changePageLeft}>
-          <Button type="button" text="Vamos lá" />
+          <Button type="button" text="próximo" />
         </Grid>
       </Grid>
     </AnimationContainer>
