+++ src/Components/Form/styles.ts
@@ -1,0 +1,15 @@
+import styled, { keyframes } from 'styled-components';
+
+const appearFromRight = keyframes`
+from {
+  opacity: 0;
+  transform: translateX(50px);
+}
+to {
+  opacity: 1;
+  transform: translateX(0);
+}
+`;
+export const AnimationContainer = styled.div`
+  animation: ${appearFromRight} 1s;
+`;
