+++ src/pages/Forms/styles.ts
@@ -1,4 +1,4 @@
-import styled, { keyframes } from 'styled-components';
+import styled from 'styled-components';
 import LogoLeves from '../../assets/logoleves.png';
 // import { shade } from 'polished';
 
@@ -27,20 +27,6 @@ export const Content = styled.div`
     margin-left: 5em;
     margin-bottom: 0.1em;
   }
-`;
-
-const appearFromRight = keyframes`
-  from {
-    opacity: 0;
-    transform: translateX(50px);
-  }
-  to {
-    opacity: 1;
-    transform: translateX(0);
-  }
-`;
-export const AnimationContainer = styled.div`
-  animation: ${appearFromRight} 1s;
 `;
 
 export const Background = styled.div`
